Properly namespace annotation management to support multiple active InfoPanels
Had to extend the plugin base class interface to access the plugin's namespaced annotate module as well as making the annotate module's #hide() and #show() methods use its own namespacing. This shouldn't break much compatibility as tweaking a plugin/the info panel to be compatible with bookmarklet style plugins is very minor.